backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
requests==2.31.0
python-dotenv==1.0.0
googlemaps==4.10.0
httpx==0.26.0
aiohttp==3.9.3
beautifulsoup4==4.12.3
redis==5.0.1
pymongo==4.6.1
google-generativeai==0.3.2
alembic==1.13.1
celery==5.3.6
gunicorn==21.2.0
# Hunter module dependencies
pdfplumber==0.10.4
dnspython==2.5.0
pdf2image==1.17.0
pytesseract==0.3.10
lxml==5.1.0
urllib3==2.1.0
